Move Github PR class to github file.
